setting the Driver

+++ Es3/src/main/java/it/polito/bigdata/hadoop/DriverBigData.java
@@ -3,6 +3,7 @@ package it.polito.bigdata.hadoop;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
@@ -65,7 +66,7 @@ implements Tool {
     
     // Set map output key and value classes
     job.setMapOutputKeyClass(Text.class);
-    job.setMapOutputValueClass(IntWritable.class);
+    job.setMapOutputValueClass(DoubleWritable.class);
     
     // Set reduce class
     job.setReducerClass(ReducerBigData.class);
